Use a rate-limiter to back off on DynamoDB throttling
By using a rate-limiter we can slow down in proportion to the amount
we are being throttled - DynamoDB may only reject one or two items out
of a batch but previously we were not slowing down until the entire
batch was rejected.

DynamoDB internal metrics refer to "Throttled events" being part of a
batch and "Throttled requests" meaning the entire batch was rejected.

ThrottleLimit effectively becomes the rate at which "Throttled events"
will be seen when under load, per process.

Signed-off-by: Bryan Boreham <bryan@weave.works>